feat(connect-wallet-modal): reorganize components, improve styles, rework scroll
