+++ packages/connect-wallet-modal/src/components/ConnectButton/index.ts
@@ -1,0 +1,1 @@
+export { default as ConnectButton } from './ConnectButton';
