+++ packages/connect-wallet-modal/src/components/Terms/Terms.tsx
