+++ packages/connect-wallet-modal/src/components/Terms/index.ts
