+++ packages/connect-wallet-modal/src/components/Terms/styles.tsx
