+++ packages/connect-wallet-modal/src/components/WalletsModal/LidoModalLogo.tsx
