+++ packages/connect-wallet-modal/src/components/WalletsModalForEth/types.ts
@@ -1,5 +1,5 @@
 import { WalletsModalProps } from '../WalletsModal';
-import { WalletId } from '../constants';
+import { WalletId } from '../../constants';
 
 export type WalletsModalForEthProps = Omit<WalletsModalProps, 'children'> & {
   hiddenWallets?: WalletId[];
