+++ packages/connect-wallet-modal/src/components/index.ts
@@ -1,0 +1,4 @@
+export * from './WalletsModal';
+export * from './WalletsModalForEth';
+export * from './Terms';
+export * from './ConnectButton';
