+++ packages/connect-wallet-modal/src/connectButtons/connectBraveWallet.tsx
@@ -3,8 +3,8 @@ import { useConnectorBraveWallet, helpers } from '@reef-knot/web3-react';
 import { Brave as WalletIcon } from '@reef-knot/wallets-icons/react';
 import { CONFLICTS } from '../constants/conflictChecks';
 import { ConnectWalletProps } from './types';
-import ConnectButton from './connectButton';
-import checkConflicts from './checkConflicts';
+import { ConnectButton } from '../components';
+import checkConflicts from '../helpers/checkConflicts';
 import { checkIfBraveBrowser } from '../helpers';
 
 const ConnectBraveWallet: FC<ConnectWalletProps> = (
