+++ packages/connect-wallet-modal/src/connectButtons/connectImToken.tsx
@@ -2,7 +2,7 @@ import { FC, useCallback } from 'react';
 import { useConnectorImToken } from '@reef-knot/web3-react';
 import { ImToken as WalletIcon } from '@reef-knot/wallets-icons/react';
 import { ConnectWalletProps } from './types';
-import ConnectButton from './connectButton';
+import { ConnectButton } from '../components';
 import { isIOS, isAndroid } from '../helpers';
 
 const ConnectImToken: FC<ConnectWalletProps> = (props) => {
