+++ packages/connect-wallet-modal/src/connectButtons/connectMathWallet.tsx
@@ -5,9 +5,9 @@ import {
 } from '@reef-knot/wallets-icons/react';
 import { useConnectorMathWallet } from '@reef-knot/web3-react';
 import { ConnectWalletProps } from './types';
-import ConnectButton from './connectButton';
+import { ConnectButton } from '../components';
 import { CONFLICTS } from '../constants/conflictChecks';
-import checkConflicts from './checkConflicts';
+import checkConflicts from '../helpers/checkConflicts';
 
 const ConnectMathWallet: FC<ConnectWalletProps> = (
   props: ConnectWalletProps,
