+++ packages/connect-wallet-modal/src/connectButtons/connectMetamask.tsx
@@ -6,8 +6,8 @@ import {
 } from '@reef-knot/wallets-icons/react';
 import { CONFLICTS } from '../constants/conflictChecks';
 import { ConnectWalletProps } from './types';
-import ConnectButton from './connectButton';
-import checkConflicts from './checkConflicts';
+import { ConnectButton } from '../components';
+import checkConflicts from '../helpers/checkConflicts';
 
 const ConnectMetamask: FC<ConnectWalletProps> = (props: ConnectWalletProps) => {
   const {
