+++ packages/connect-wallet-modal/src/connectButtons/connectOperaWallet.tsx
@@ -2,7 +2,7 @@ import { FC, useCallback } from 'react';
 import { useConnectorOperaWallet } from '@reef-knot/web3-react';
 import { OperaWallet as WalletIcon } from '@reef-knot/wallets-icons/react';
 import { ConnectWalletProps } from './types';
-import ConnectButton from './connectButton';
+import { ConnectButton } from '../components';
 
 const ConnectOperaWallet: FC<ConnectWalletProps> = (
   props: ConnectWalletProps,
