+++ packages/connect-wallet-modal/src/connectButtons/connectTally.tsx
@@ -2,7 +2,7 @@ import { FC, useCallback } from 'react';
 import { useConnectorTally } from '@reef-knot/web3-react';
 import { Tally } from '@reef-knot/wallets-icons/react';
 import { ConnectWalletProps } from './types';
-import ConnectButton from './connectButton';
+import { ConnectButton } from '../components';
 import { isMobileOrTablet } from '../helpers';
 
 const ConnectTally: FC<ConnectWalletProps> = (props: ConnectWalletProps) => {
