+++ packages/connect-wallet-modal/src/connectButtons/connectZenGo.tsx
@@ -5,7 +5,7 @@ import {
 } from '@reef-knot/web3-react';
 import { ZenGo as WalletIcon } from '@reef-knot/wallets-icons/react';
 import { ConnectWalletProps } from './types';
-import ConnectButton from './connectButton';
+import { ConnectButton } from '../components';
 import { isMobileOrTablet } from '../helpers';
 
 let newWindow: Window | null = null;
