+++ packages/connect-wallet-modal/src/index.ts
@@ -1,4 +1,2 @@
-export * from './WalletsModal';
-export * from './WalletsModalForEth';
-export * from './Terms';
+export * from './components';
 export * from './connectButtons';
